+++ erigon-lib/chain/chain_config_test.go
@@ -66,3 +66,14 @@ func TestConfigValueLookup(t *testing.T) {
 	assert.Equal(t, ConfigValueLookup(burntContract, 41874000), address2)
 	assert.Equal(t, ConfigValueLookup(burntContract, 41874000+1), address2)
 }
+
+// Check that EIP-4844 values are returned for a nil BlobSchedule
+func TestNilBlobSchedule(t *testing.T) {
+	var b *BlobSchedule
+	assert.Equal(t, uint64(3), b.TargetBlobsPerBlock(false /* isPrague */))
+	assert.Equal(t, uint64(6), b.MaxBlobsPerBlock(false /* isPrague */))
+	assert.Equal(t, uint64(3338477), b.BaseFeeUpdateFraction(false /* isPrague */))
+	assert.Equal(t, uint64(3), b.TargetBlobsPerBlock(true /* isPrague */))
+	assert.Equal(t, uint64(6), b.MaxBlobsPerBlock(true /* isPrague */))
+	assert.Equal(t, uint64(3338477), b.BaseFeeUpdateFraction(true /* isPrague */))
+}
